tutorials/openai_time_demo/requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0  # For better console output
